added implementation of ImageRepository
The image repository has now an implemtation of a fetchImage and a storeImage. Both functions were not tested yet, so we should test them before merge. Next thing to do is to set a default picture for each user, and add a function to give possibility of user to change the profile picture. I also added two functions to the profileViewModel to be able to store an image.